requirements.txt
Django==2.2.8
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
django-allauth==0.41.0
django-rest-auth==0.9.5
django-rest-swagger==2.2.0

faker==4.0.0

pre-commit==1.21.0
